feat(js): separate notification action concern
