+++ packages/js/src/ui/components/Notification/index.ts
@@ -1,2 +1,3 @@
 export * from './Notification';
 export { NotificationList } from './NotificationList';
+export * from './NotificationActions';
